pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audifonospro"
version = "0.1.0"
description = "Multi-device personal audio system — translation, ANC, EQ, cinema mode"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.60.0",
    "dasbus>=1.6.0",
    "sounddevice>=0.5.0",
    "numpy>=2.0.0",
    "scipy>=1.13.0",
    "noisereduce>=3.0.0",
    "edge-tts>=6.1.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "bleak>=0.21.0",
    "openai>=1.30.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
audifonospro = "audifonospro.main:main"

[tool.hatch.build.targets.wheel]
packages = ["audifonospro"]
